+++ custom-elements-manifest.config.mjs
@@ -1,4 +1,5 @@
 import fs from 'fs';
+import path from 'path';
 import sortItems from './cem/sort-items.js';
 import removePrivateMembers from './cem/remove-private-members.js';
 import identifyReadonlyMembers from './cem/identify-readonly-members.js';
@@ -14,9 +15,15 @@ try {
 catch (e) {
 }
 
+// For our WDS plugin "cem-analyzer-plugin", we need to generate the CEM to a temporary dir.
+// The CEM analyzer CLI doesn't have a command line param to override the outdir so we're using this env var hack.
+const outdir = (process.env.CEM_OUT_DIR != null)
+  ? path.relative(process.cwd(), process.env.CEM_OUT_DIR)
+  : 'dist';
+
 export default {
   globs: ['src/**/cc-*.js'],
-  outdir: 'dist',
+  outdir,
   // dev: true,
   // watch: true,
   plugins: [